start optimisation - separate generating from suggested data structures
