+++ src/Nethermind/Nethermind.Consensus/Processing/BlockProcessor.BlockValidationTransactionsExecutor.cs
@@ -64,7 +64,7 @@ namespace Nethermind.Consensus.Processing
 
                 for (int i = 0; i < block.Transactions.Length; i++)
                 {
-                    _balBuilder?.GeneratedBlockAccessList.IncrementBlockAccessIndex();
+                    // _balBuilder?.GeneratedBlockAccessList.IncrementBlockAccessIndex();
                     Transaction currentTx = block.Transactions[i];
                     ProcessTransaction(block, currentTx, i, receiptsTracer, processingOptions);
 
@@ -74,7 +74,7 @@ namespace Nethermind.Consensus.Processing
                         _balBuilder?.ValidateBlockAccessList(block.Header, (ushort)(i + 1), gasRemaining!.Value);
                     }
                 }
-                _balBuilder?.GeneratedBlockAccessList.IncrementBlockAccessIndex();
+                // _balBuilder?.GeneratedBlockAccessList.IncrementBlockAccessIndex();
 
                 return [.. receiptsTracer.TxReceipts];
             }
