+++ src/Nethermind/Nethermind.Core/BlockAccessLists/BlockAccessList.cs
@@ -7,18 +7,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Text.Json.Serialization;
-using Nethermind.Int256;
 
 [assembly: InternalsVisibleTo("Nethermind.Core.Test")]
 [assembly: InternalsVisibleTo("Nethermind.State.Test")]
 
 namespace Nethermind.Core.BlockAccessLists;
 
-public class BlockAccessList : IEquatable<BlockAccessList>, IJournal<int>
+public class BlockAccessList : IEquatable<BlockAccessList>
 {
-    [JsonIgnore]
-    public int Index = 0;
-
     /// storage keys across all accounts + addresses
     [JsonIgnore]
     public int ItemCount { get; set; }
@@ -27,7 +23,6 @@ public class BlockAccessList : IEquatable<BlockAccessList>, IJournal<int>
     public bool HasAccount(Address address) => _accountChanges.ContainsKey(address);
 
     private readonly SortedDictionary<Address, AccountChanges> _accountChanges = [];
-    private readonly Stack<Change> _changes = new();
 
     public BlockAccessList()
     {
@@ -53,296 +48,7 @@ public class BlockAccessList : IEquatable<BlockAccessList>, IJournal<int>
     public static bool operator !=(BlockAccessList left, BlockAccessList right) =>
         !(left == right);
 
-    public void Merge(BlockAccessList other)
-    {
-        foreach (AccountChanges otherAccountChange in other.AccountChanges)
-        {
-            if (_accountChanges.TryGetValue(otherAccountChange.Address, out AccountChanges? accountChange))
-            {
-                accountChange.Merge(otherAccountChange);
-            }
-            else
-            {
-                _accountChanges.Add(otherAccountChange.Address, otherAccountChange);
-            }
-        }
-    }
-
     public AccountChanges? GetAccountChanges(Address address) => _accountChanges.TryGetValue(address, out AccountChanges? value) ? value : null;
-
-    public void IncrementBlockAccessIndex()
-    {
-        _changes.Clear();
-        Index++;
-    }
-
-    public void RollbackCurrentIndex()
-    {
-        Restore(0);
-        _changes.Clear();
-        Index--;
-    }
-
-    public void Clear()
-    {
-        _accountChanges.Clear();
-        _changes.Clear();
-        Index = 0;
-    }
-
-    public void AddBalanceChange(Address address, UInt256 before, UInt256 after)
-    {
-        bool isZeroBalanceChange = before == after;
-        if (address == Address.SystemUser && isZeroBalanceChange)
-        {
-            return;
-        }
-
-        AccountChanges accountChanges = GetOrAddAccountChanges(address);
-
-        // don't add zero balance transfers, but add empty account changes
-        if (isZeroBalanceChange)
-        {
-            return;
-        }
-
-        bool changedDuringTx = HasBalanceChangedDuringTx(address, before, after);
-        accountChanges.PopBalanceChange(Index, out BalanceChange? oldBalanceChange);
-
-        _changes.Push(new()
-        {
-            Address = address,
-            Type = ChangeType.BalanceChange,
-            PreviousValue = oldBalanceChange,
-            PreTxBalance = before,
-            BlockAccessIndex = Index
-        });
-
-        if (changedDuringTx)
-        {
-            accountChanges.AddBalanceChange(new(Index, after));
-        }
-    }
-
-    public void AddCodeChange(Address address, byte[] before, ReadOnlyMemory<byte> after)
-    {
-        AccountChanges accountChanges = GetOrAddAccountChanges(address);
-
-        if (before.AsSpan().SequenceEqual(after.Span))
-        {
-            return;
-        }
-
-        bool changedDuringTx = HasCodeChangedDuringTx(accountChanges.Address, before, after.Span);
-        accountChanges.PopCodeChange(Index, out CodeChange? oldCodeChange);
-        _changes.Push(new()
-        {
-            Address = address,
-            Type = ChangeType.CodeChange,
-            PreviousValue = oldCodeChange,
-            PreTxCode = before
-            // N.B. don't need PreTxCode as SELFDESTRUCT cannot be first code change of tx
-        });
-
-        if (changedDuringTx)
-        {
-            accountChanges.AddCodeChange(new(Index, after.ToArray()));
-        }
-    }
-
-    public void AddNonceChange(Address address, ulong newNonce)
-    {
-        if (newNonce == 0)
-        {
-            return;
-        }
-
-        AccountChanges accountChanges = GetOrAddAccountChanges(address);
-
-        accountChanges.PopNonceChange(Index, out NonceChange? oldNonceChange);
-        _changes.Push(new()
-        {
-            Address = address,
-            Type = ChangeType.NonceChange,
-            PreviousValue = oldNonceChange
-        });
-
-        accountChanges.AddNonceChange(new(Index, newNonce));
-    }
-
-    public void AddAccountRead(Address address)
-    {
-        if (!_accountChanges.ContainsKey(address))
-        {
-            _accountChanges.Add(address, new(address));
-        }
-    }
-
-    public void AddStorageChange(Address address, UInt256 key, UInt256 before, UInt256 after)
-    {
-        AccountChanges accountChanges = GetOrAddAccountChanges(address);
-
-        if (before != after)
-        {
-            StorageChange(accountChanges, key, before, after);
-        }
-    }
-
-    public void AddStorageChange(in StorageCell storageCell, UInt256 before, UInt256 after)
-        => AddStorageChange(storageCell.Address, storageCell.Index, before, after);
-
-    public void AddStorageRead(in StorageCell storageCell) =>
-        AddStorageRead(storageCell.Address, storageCell.Index);
-
-    public void AddStorageRead(Address address, UInt256 key)
-    {
-        AccountChanges accountChanges = GetOrAddAccountChanges(address);
-
-        if (!accountChanges.HasStorageChange(key))
-        {
-            accountChanges.AddStorageRead(key);
-        }
-    }
-
-    public void DeleteAccount(Address address, UInt256 oldBalance)
-    {
-        AccountChanges accountChanges = GetOrAddAccountChanges(address);
-
-        // Push revertible changes for each storage change that will be cleared.
-        // Push ALL changes per slot in reverse order so they restore in correct order (LIFO).
-        foreach (SlotChanges slotChanges in accountChanges.StorageChanges)
-        {
-            // Push changes in reverse order so they restore in original order
-            foreach (KeyValuePair<int, StorageChange> change in slotChanges.Changes)
-            {
-                _changes.Push(new()
-                {
-                    Address = address,
-                    Type = ChangeType.StorageChange,
-                    Slot = slotChanges.Slot,
-                    PreviousValue = change.Value,
-                    BlockAccessIndex = Index
-                });
-            }
-        }
-
-        // Push revertible changes for nonce changes (reverse order for correct restore)
-        IList<NonceChange> nonceChanges = accountChanges.NonceChanges;
-        int nonceCount = nonceChanges.Count;
-        for (int i = nonceCount - 1; i >= 0; i--)
-        {
-            _changes.Push(new()
-            {
-                Address = address,
-                Type = ChangeType.NonceChange,
-                PreviousValue = nonceChanges[i],
-                BlockAccessIndex = Index
-            });
-        }
-
-        // Push revertible changes for code changes (reverse order for correct restore)
-        IList<CodeChange> codeChanges = accountChanges.CodeChanges;
-        int codeCount = codeChanges.Count;
-        for (int i = codeCount - 1; i >= 0; i--)
-        {
-            _changes.Push(new()
-            {
-                Address = address,
-                Type = ChangeType.CodeChange,
-                PreviousValue = codeChanges[i],
-                BlockAccessIndex = Index
-            });
-        }
-
-        accountChanges.SelfDestruct();
-        AddBalanceChange(address, oldBalance, 0);
-    }
-
-    private void StorageChange(AccountChanges accountChanges, in UInt256 key, in UInt256 before, in UInt256 after)
-    {
-        SlotChanges slotChanges = accountChanges.GetOrAddSlotChanges(key);
-
-        bool changedDuringTx = HasStorageChangedDuringTx(accountChanges.Address, key, before, after);
-        slotChanges.TryPopStorageChange(Index, out StorageChange? oldStorageChange);
-
-        _changes.Push(new()
-        {
-            Address = accountChanges.Address,
-            BlockAccessIndex = Index,
-            Slot = key,
-            Type = ChangeType.StorageChange,
-            PreviousValue = oldStorageChange,
-            PreTxStorage = before
-        });
-
-        if (changedDuringTx)
-        {
-            slotChanges.AddStorageChange(new(Index, after));
-            accountChanges.RemoveStorageRead(key);
-        }
-        else
-        {
-            accountChanges.ClearEmptySlotChangesAndAddRead(key);
-        }
-    }
-
-    public int TakeSnapshot()
-        => _changes.Count;
-
-    public void Restore(int snapshot, int? blockAccessIndex = null)
-    {
-        snapshot = int.Max(0, snapshot);
-        while (_changes.Count > snapshot)
-        {
-            Change change = _changes.Pop();
-            AccountChanges accountChanges = _accountChanges[change.Address];
-            switch (change.Type)
-            {
-                case ChangeType.BalanceChange:
-                    BalanceChange? previousBalance = change.PreviousValue is null ? null : (BalanceChange)change.PreviousValue;
-
-                    // balance could have gone back to pre-tx value
-                    // so would already be empty
-                    accountChanges.PopBalanceChange(change.BlockAccessIndex, out _); // todo: this index must be the same?
-                    if (previousBalance is not null)
-                    {
-                        accountChanges.AddBalanceChange(previousBalance.Value);
-                    }
-                    break;
-                case ChangeType.CodeChange:
-                    CodeChange? previousCode = change.PreviousValue is null ? null : (CodeChange)change.PreviousValue;
-
-                    accountChanges.PopCodeChange(Index, out _);
-                    if (previousCode is not null)
-                    {
-                        accountChanges.AddCodeChange(previousCode.Value);
-                    }
-                    break;
-                case ChangeType.NonceChange:
-                    NonceChange? previousNonce = change.PreviousValue is null ? null : (NonceChange)change.PreviousValue;
-
-                    accountChanges.PopNonceChange(Index, out _);
-                    if (previousNonce is not null)
-                    {
-                        accountChanges.AddNonceChange(previousNonce.Value);
-                    }
-                    break;
-                case ChangeType.StorageChange:
-                    StorageChange? previousStorage = change.PreviousValue is null ? null : (StorageChange)change.PreviousValue;
-                    SlotChanges slotChanges = accountChanges.GetOrAddSlotChanges(change.Slot!.Value);
-
-                    slotChanges.TryPopStorageChange(Index, out _);
-                    if (previousStorage is not null)
-                    {
-                        slotChanges.AddStorageChange(previousStorage.Value);
-                        accountChanges.RemoveStorageRead(change.Slot.Value);
-                    }
-
-                    accountChanges.ClearEmptySlotChangesAndAddRead(change.Slot!.Value);
-                    break;
-            }
-        }
-    }
 
     public IEnumerable<ChangeAtIndex> GetChangesAtIndex(ushort index)
     {
@@ -367,7 +73,7 @@ public class BlockAccessList : IEquatable<BlockAccessList>, IJournal<int>
     public override string ToString()
     {
         StringBuilder sb = new();
-        sb.AppendLine($"BlockAccessList (Index={Index}, Accounts={_accountChanges.Count})");
+        sb.AppendLine($"BlockAccessList (Accounts={_accountChanges.Count})");
         foreach (AccountChanges ac in _accountChanges.Values)
         {
             sb.AppendLine($"  {ac}");
@@ -390,151 +96,6 @@ public class BlockAccessList : IEquatable<BlockAccessList>, IJournal<int>
         {
             _accountChanges.Remove(address);
         }
-    }
-
-    private bool HasBalanceChangedDuringTx(Address address, UInt256 beforeInstr, UInt256 afterInstr)
-    {
-        AccountChanges accountChanges = _accountChanges[address];
-        IList<BalanceChange> balanceChanges = accountChanges.BalanceChanges;
-        int count = balanceChanges.Count;
-
-        if (count == 0)
-        {
-            // first balance change of block
-            // return balance prior to this instruction
-            return beforeInstr != afterInstr;
-        }
-
-        for (int i = count - 1; i >= 0; i--)
-        {
-            BalanceChange balanceChange = balanceChanges[i];
-            if (balanceChange.BlockAccessIndex != Index)
-            {
-                // balance changed in previous tx in block
-                return balanceChange.PostBalance != afterInstr;
-            }
-        }
-
-        // balance only changed within this transaction
-        foreach (Change change in _changes)
-        {
-            if (change.Type == ChangeType.BalanceChange && change.Address == address && change.PreviousValue is null)
-            {
-                // first change of this transaction & block
-                return change.PreTxBalance!.Value != afterInstr;
-            }
-        }
-
-        // should never happen
-        throw new InvalidOperationException("Error calculating pre tx balance");
-    }
-
-    private bool HasStorageChangedDuringTx(Address address, UInt256 key, in UInt256 beforeInstr, in UInt256 afterInstr)
-    {
-        AccountChanges accountChanges = _accountChanges[address];
-
-        if (!accountChanges.TryGetSlotChanges(key, out SlotChanges? slotChanges) || slotChanges.Changes.Count == 0)
-        {
-            // first storage change of block
-            // return storage prior to this instruction
-            return beforeInstr != afterInstr;
-        }
-
-        IList<StorageChange> values = slotChanges.Changes.Values;
-        for (int i = values.Count - 1; i >= 0; i--)
-        {
-            StorageChange storageChange = values[i];
-            if (storageChange.BlockAccessIndex != Index)
-            {
-                // storage changed in previous tx in block
-                return storageChange.NewValue != afterInstr;
-            }
-        }
-
-        // storage only changed within this transaction
-        foreach (Change change in _changes)
-        {
-            if (
-                change.Type == ChangeType.StorageChange &&
-                change.Address == address &&
-                change.Slot == key &&
-                change.PreviousValue is null)
-            {
-                // first change of this transaction & block
-                return change.PreTxStorage is null || change.PreTxStorage != afterInstr;
-            }
-        }
-
-        // should never happen
-        throw new InvalidOperationException("Error calculating pre tx storage");
-    }
-
-    private bool HasCodeChangedDuringTx(Address address, in ReadOnlySpan<byte> beforeInstr, in ReadOnlySpan<byte> afterInstr)
-    {
-        AccountChanges accountChanges = _accountChanges[address];
-        IList<CodeChange> codeChanges = accountChanges.CodeChanges;
-        int count = codeChanges.Count;
-
-        if (count == 0)
-        {
-            // first code change of block
-            // return code prior to this instruction
-            return !beforeInstr.SequenceEqual(afterInstr);
-        }
-
-        for (int i = count - 1; i >= 0; i--)
-        {
-            CodeChange codeChange = codeChanges[i];
-            if (codeChange.BlockAccessIndex != Index)
-            {
-                // code changed in previous tx in block
-                return !codeChange.NewCode.AsSpan().SequenceEqual(afterInstr);
-            }
-        }
-
-        // code only changed within this transaction
-        foreach (Change change in _changes)
-        {
-            if (change.Type == ChangeType.CodeChange && change.Address == address && change.PreviousValue is null)
-            {
-                // first change of this transaction & block
-                return change.PreTxCode is null || !change.PreTxCode.AsSpan().SequenceEqual(afterInstr);
-            }
-        }
-
-        // should never happen
-        throw new InvalidOperationException("Error calculating pre tx code");
-    }
-
-    private AccountChanges GetOrAddAccountChanges(Address address)
-    {
-        if (!_accountChanges.TryGetValue(address, out AccountChanges? existing))
-        {
-            AccountChanges accountChanges = new(address);
-            _accountChanges.Add(address, accountChanges);
-            return accountChanges;
-        }
-        return existing;
-    }
-
-    private enum ChangeType
-    {
-        BalanceChange = 0,
-        CodeChange = 1,
-        NonceChange = 2,
-        StorageChange = 3
-    }
-
-    private readonly struct Change
-    {
-        public Address Address { get; init; }
-        public UInt256? Slot { get; init; }
-        public ChangeType Type { get; init; }
-        public IIndexedChange? PreviousValue { get; init; }
-        public UInt256? PreTxBalance { get; init; }
-        public UInt256? PreTxStorage { get; init; }
-        public byte[]? PreTxCode { get; init; }
-        public int BlockAccessIndex { get; init; }
     }
 }
